Many many updates to the docs, including a quick-start
