+++ build.sh
@@ -227,18 +227,50 @@ fi
 
 if hasArg tests || (( ${NUMARGS} == 0 )); then
     BUILD_TESTS=ON
-    COMPILE_DIST_LIBRARY=ON
-    ENABLE_NN_DEPENDENCIES=ON
-    COMPILE_NN_LIBRARY=ON
     CMAKE_TARGET="${CMAKE_TARGET};${TEST_TARGETS}"
+
+    # Force compile nn library when needed test targets are specified
+    if [[ $CMAKE_TARGET == *"CLUSTER_TEST"* || \
+          $CMAKE_TARGET == *"SPARSE_DIST_TEST"* || \
+          $CMAKE_TARGET == *"SPARSE_NEIGHBORS_TEST"* || \
+          $CMAKE_TARGET == *"NEIGHBORS_TEST"* || \
+          $CMAKE_TARGET == *"STATS_TEST"* ]]; then
+      echo "-- Enabling nearest neighbors lib for gtests"
+      ENABLE_NN_DEPENDENCIES=ON
+      COMPILE_NN_LIBRARY=ON
+    fi
+
+    # Force compile distance library when needed test targets are specified
+    if [[ $CMAKE_TARGET == *"CLUSTER_TEST"* || \
+          $CMAKE_TARGET == *"DISTANCE_TEST"* || \
+          $CMAKE_TARGET == *"SPARSE_DIST_TEST" || \
+          $CMAKE_TARGET == *"SPARSE_NEIGHBORS_TEST"* || \
+          $CMAKE_TARGET == *"NEIGHBORS_TEST" || \
+          $CMAKE_TARGET == *"STATS_TEST"* ]]; then
+      echo "-- Enabling distance lib for gtests"
+      COMPILE_DIST_LIBRARY=ON
+    fi
 fi
 
 if hasArg bench || (( ${NUMARGS} == 0 )); then
     BUILD_BENCH=ON
-    COMPILE_DIST_LIBRARY=ON
-    ENABLE_NN_DEPENDENCIES=ON
-    COMPILE_NN_LIBRARY=ON
     CMAKE_TARGET="${CMAKE_TARGET};${BENCH_TARGETS}"
+
+    # Force compile nn library when needed benchmark targets are specified
+    if [[ $CMAKE_TARGET == *"CLUSTER_BENCH"* || \
+          $CMAKE_TARGET == *"NEIGHBORS_BENCH"*  ]]; then
+      echo "-- Enabling nearest neighbors lib for benchmarks"
+      ENABLE_NN_DEPENDENCIES=ON
+      COMPILE_NN_LIBRARY=ON
+    fi
+
+    # Force compile distance library when needed benchmark targets are specified
+    if [[ $CMAKE_TARGET == *"CLUSTER_BENCH"* || \
+          $CMAKE_TARGET == *"NEIGHBORS_BENCH"* ]]; then
+      echo "-- Enabling distance lib for benchmarks"
+      COMPILE_DIST_LIBRARY=ON
+    fi
+
 fi
 
 if hasArg --buildfaiss; then
