+++ cpp/include/raft/neighbors/brute_force.cuh
@@ -23,26 +23,52 @@
 namespace raft::neighbors::brute_force {
 
 /**
- * @brief Performs a k-select across row partitioned index/distance
+ * @brief Performs a k-select across several (contiguous) row-partitioned index/distance
  * matrices formatted like the following:
- * row1: k0, k1, k2
- * row2: k0, k1, k2
- * row3: k0, k1, k2
- * row1: k0, k1, k2
- * row2: k0, k1, k2
- * row3: k0, k1, k2
  *
+ * part1row1: k0, k1, k2, k3
+ * part1row2: k0, k1, k2, k3
+ * part1row3: k0, k1, k2, k3
+ * part2row1: k0, k1, k2, k3
+ * part2row2: k0, k1, k2, k3
+ * part2row3: k0, k1, k2, k3
  * etc...
+ *
+ * The example above shows what an aggregated index/distance matrix
+ * would look like with two partitions when n_samples=3 and k=4.
+ *
+ * When working with extremely large data sets that have been broken
+ * over multiple indexes, such as when computing over multiple GPUs,
+ * the ids will often start at 0 for each local knn index but the
+ * global ids need to be used when merging them together. An optional
+ * translations vector can be supplied to map the starting id of
+ * each partition to its global id so that the final merged knn
+ * is based on the global ids.
+ *
+ * Usage example:
+ * @code{.cpp}
+ *  #include <raft/core/handle.hpp>
+ *  #include <raft/neighbors/brute_force.cuh>
+ *  using namespace raft::neighbors;
+ *
+ *  raft::handle_t handle;
+ *  ...
+ *  compute multiple knn graphs and aggregate row-wise
+ *  (see detailed description above)
+ *  ...
+ *  brute_force::knn_merge_parts(handle, in_keys, in_values, out_keys, out_values, n_samples);
+ * @endcode
  *
  * @tparam idx_t
  * @tparam value_t
+ *
  * @param[in] handle
  * @param[in] in_keys matrix of input keys (size n_samples * n_parts * k)
  * @param[in] in_values matrix of input values (size n_samples * n_parts * k)
  * @param[out] out_keys matrix of output keys (size n_samples * k)
  * @param[out] out_values matrix of output values (size n_samples * k)
- * @param[in] n_samples number of rows in each part
- * @param[in] translations optional vector of starting index mappings for each partition
+ * @param[in] n_samples number of rows in each partition
+ * @param[in] translations optional vector of starting global id mappings for each local partition
  */
 template <typename idx_t, typename value_t>
 inline void knn_merge_parts(
@@ -81,17 +107,31 @@ inline void knn_merge_parts(
  * row- or column-major but the output matrices will always be in
  * row-major format.
  *
- * @param[in] handle the cuml handle to use
- * @param[in] index vector of device matrices (each size m_i*d) to be used as the knn index
- * @param[in] search matrix (size n*d) to be used for searching the index
- * @param[out] indices matrix (size n*k) to store output knn indices
- * @param[out] distances matrix (size n*k) to store the output knn distance
- * @param[in] k the number of nearest neighbors to return
- * @param[in] metric distance metric to use. Euclidean (L2) is used by default
- * @param[in] metric_arg the value of `p` for Minkowski (l-p) distances. This
+ * Usage example:
+ * @code{.cpp}
+ *  #include <raft/core/handle.hpp>
+ *  #include <raft/neighbors/brute_force.cuh>
+ *  #include <raft/distance/distance_types.hpp>
+ *  using namespace raft::neighbors;
+ *
+ *  raft::handle_t handle;
+ *  ...
+ *  int k = 10;
+ *  auto metric = raft::distance::DistanceType::L2SqrtExpanded;
+ *  brute_force::knn(handle, index, search, indices, distances, k, metric);
+ * @endcode
+ *
+ * @param[in] handle: the cuml handle to use
+ * @param[in] index: vector of device matrices (each size m_i*d) to be used as the knn index
+ * @param[in] search: matrix (size n*d) to be used for searching the index
+ * @param[out] indices: matrix (size n*k) to store output knn indices
+ * @param[out] distances: matrix (size n*k) to store the output knn distance
+ * @param[in] k: the number of nearest neighbors to return
+ * @param[in] metric: distance metric to use. Euclidean (L2) is used by default
+ * @param[in] metric_arg: the value of `p` for Minkowski (l-p) distances. This
  * 					 is ignored if the metric_type is not Minkowski.
- * @param[in] translations starting offsets for partitions. should be the same size
- *            as input vector.
+ * @param[in] global_id_offset: optional starting global id mapping for the local partition
+ *                              (assumes the index contains contiguous ids in the global id space)
  */
 template <typename idx_t,
           typename value_t,
@@ -105,9 +145,9 @@ void knn(raft::handle_t const& handle,
          raft::device_matrix_view<idx_t, matrix_idx, row_major> indices,
          raft::device_matrix_view<value_t, matrix_idx, row_major> distances,
          value_int k,
-         distance::DistanceType metric                  = distance::DistanceType::L2Unexpanded,
-         std::optional<float> metric_arg                = std::make_optional<float>(2.0f),
-         std::optional<std::vector<idx_t>> translations = std::nullopt)
+         distance::DistanceType metric         = distance::DistanceType::L2Unexpanded,
+         std::optional<float> metric_arg       = std::make_optional<float>(2.0f),
+         std::optional<idx_t> global_id_offset = std::nullopt)
 {
   RAFT_EXPECTS(index[0].extent(1) == search.extent(1),
                "Number of dimensions for both index and search matrices must be equal");
@@ -129,7 +169,10 @@ void knn(raft::handle_t const& handle,
     sizes.push_back(index[i].extent(0));
   }
 
-  std::vector<idx_t>* trans = translations.has_value() ? &(*translations) : nullptr;
+  std::vector<idx_t> trans;
+  if (global_id_offset.has_value()) { trans.push_back(global_id_offset.value()); }
+
+  std::vector<idx_t>* trans_arg = global_id_offset.has_value() ? &trans : nullptr;
 
   raft::spatial::knn::detail::brute_force_knn_impl(handle,
                                                    inputs,
@@ -143,7 +186,7 @@ void knn(raft::handle_t const& handle,
                                                    k,
                                                    rowMajorIndex,
                                                    rowMajorQuery,
-                                                   trans,
+                                                   trans_arg,
                                                    metric,
                                                    metric_arg.value_or(2.0f));
 }
