+++ cpp/include/raft/spatial/knn/ivf_flat.cuh
@@ -33,7 +33,6 @@
 namespace raft::spatial::knn::ivf_flat {
 
 using raft::neighbors::ivf_flat::build;
-using raft::neighbors::ivf_flat::build_index;
 using raft::neighbors::ivf_flat::extend;
 using raft::neighbors::ivf_flat::search;
