+++ cpp/test/neighbors/knn.cu
@@ -94,7 +94,8 @@ class KNNTest : public ::testing::TestWithParam<KNNInputs> {
     auto distances =
       raft::make_device_matrix_view<T, IdxT, row_major>(distances_.data(), rows_, k_);
 
-    knn(handle, index, search, indices, distances, k_);
+    auto metric = raft::distance::DistanceType::L2Unexpanded;
+    knn(handle, index, search, indices, distances, k_, metric, std::make_optional<IdxT>(0));
 
     build_actual_output<<<raft::ceildiv(rows_ * k_, 32), 32, 0, stream>>>(
       actual_labels_.data(), rows_, k_, search_labels_.data(), indices_.data());
