+++ docs/source/cpp_api.rst
@@ -1,6 +1,7 @@
-~~~~~~~~~~~~~~~~~~~~~~
-RAFT C++ API Reference
-~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~
+C++ API Reference
+~~~~~~~~~~~~~~~~~
+
 
 
 .. _api:
