+++ docs/source/cpp_api/cluster.rst
@@ -1,7 +1,8 @@
 Cluster
 =======
 
-This page provides C++ class references for the publicly-exposed elements of the cluster package.
+This page provides C++ class references for the publicly-exposed elements of the `raft/cluster` headers. RAFT provides
+fundamental clustering algorithms which are, themselves, considered reusable building blocks for other algorithms.
 
 K-Means
 -------
