+++ docs/source/cpp_api/distance.rst
@@ -1,7 +1,8 @@
 Distance
 ========
 
-This page provides C++ class references for the publicly-exposed elements of the distance package.
+This page provides C++ class references for the publicly-exposed elements of the `raft/distance` package. RAFT's
+distances have been highly optimized and support a wide assortment of different distance measures.
 
 Distance
 ########
