+++ docs/source/cpp_api/linalg.rst
@@ -1,7 +1,10 @@
 Linear Algebra
 ==============
 
-This page provides C++ class references for the publicly-exposed elements of the (dense) linear algebra package.
+This page provides C++ class references for the publicly-exposed elements of the `raft/linalg` (dense) linear algebra headers.
+In addition to providing highly optimized arithmetic and matrix/vector operations, RAFT provides a consistent user experience
+by providing common BLAS routines, standard linear system solvers, factorization and eigenvalue solvers. Some of these routines
+hide the complexities of lower-level C-based libraries provided in the CUDA toolkit 
 
 .. doxygennamespace:: raft::linalg
     :project: RAFT
