+++ docs/source/cpp_api/matrix.rst
@@ -1,7 +1,8 @@
 Matrix
 ======
 
-This page provides C++ class references for the publicly-exposed elements of the matrix package.
+This page provides C++ class references for the publicly-exposed elements of the `raft/matrix` headers. The `raft/matrix`
+headers cover many operations on matrices that are otherwise not covered by `raft/linalg`.
 
 .. doxygennamespace:: raft::matrix
     :project: RAFT
